+++ scripts/lichessTools-init.js
@@ -32,7 +32,6 @@ lichessTools.loadTool(LiChessTools.Tools.ShapeRank);
 lichessTools.loadTool(LiChessTools.Tools.ShowOpening);
 lichessTools.loadTool(LiChessTools.Tools.PreviousGameMenu);
 lichessTools.loadTool(LiChessTools.Tools.OpeningExplorerUsers);
-lichessTools.loadTool(LiChessTools.Tools.TvGameLinkAndBookmark);
 lichessTools.loadTool(LiChessTools.Tools.ChapterNavigation);
 lichessTools.loadTool(LiChessTools.Tools.StickyPreview);
 lichessTools.loadTool(LiChessTools.Tools.StickyAnalysis);
@@ -49,7 +48,6 @@ lichessTools.loadTool(LiChessTools.Tools.CustomChatButtons);
 lichessTools.loadTool(LiChessTools.Tools.AdditionalGlyphs);
 lichessTools.loadTool(LiChessTools.Tools.ExtraChart);
 lichessTools.loadTool(LiChessTools.Tools.KeyboardHelp);
-lichessTools.loadTool(LiChessTools.Tools.FriendsTv);
 lichessTools.loadTool(LiChessTools.Tools.PreviousStudyMenu);
 lichessTools.loadTool(LiChessTools.Tools.TimelineNotify);
 lichessTools.loadTool(LiChessTools.Tools.ExplorerEval);
@@ -59,16 +57,15 @@ lichessTools.loadTool(LiChessTools.Tools.StickyStudySettings);
 lichessTools.loadTool(LiChessTools.Tools.ChapterInsert);
 lichessTools.loadTool(LiChessTools.Tools.ChapterPgnArea);
 lichessTools.loadTool(LiChessTools.Tools.MoveListOptions);
-lichessTools.loadTool(LiChessTools.Tools.StreamerTv);
 lichessTools.loadTool(LiChessTools.Tools.WikiFen);
 lichessTools.loadTool(LiChessTools.Tools.TranspositionArrows);
+lichessTools.loadTool(LiChessTools.Tools.TvOptions);
+lichessTools.loadTool(LiChessTools.Tools.ShowPawnStructure);
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);
 lichessTools.loadTool(LiChessTools.Tools.TrapValueCommand);
 lichessTools.loadTool(LiChessTools.Tools.CopyPgnCommand);
 lichessTools.loadTool(LiChessTools.Tools.ReadGameCommand);
-
-lichessTools.loadTool(LiChessTools.Tools.UserTvHistory);
 
 lichessTools.loadTool(LiChessTools.Tools.Preferences);