+++ tools/customEngineLevel/tool.js
@@ -10,10 +10,10 @@
         advanced: true
       },
       {
-        name:'noCloud',
+        name:'customEngineOptions',
         category: 'analysis',
-        type:'single',
-        possibleValues: [false,true],
+        type:'multiple',
+        possibleValues: ['noCloud','practice'],
         defaultValue: false,
         advanced: true
       }
@@ -22,88 +22,21 @@
     intl={
       'en-US':{
         'options.analysis': 'Analysis',
-        'options.customEngineLevel': 'Minimum analysis engine depth',
-        'options.noCloud': 'Always use local engine for evaluation'
+        'options.customEngineLevel': 'Custom analysis engine depth',
+        'options.customEngineOptions': 'Custom analysis engine options',
+        'customEngineOptions.noCloud': 'Always use local engine for evaluation',
+        'customEngineOptions.practice': 'Apply in Practice mode',
+        'applyInPractice': 'Custom engine settings in Practice mode'
       },
       'ro-RO':{
         'options.analysis': 'Analiz\u0103',
-        'options.customEngineLevel': 'Nivel minim pentru motorul de analiz\u0103',
-        'options.noCloud': 'Folose\u015fte numai motorul local pentru evalu\u0103ri'
+        'options.customEngineLevel': 'Nivel personalizat pentru motorul de analiz\u0103',
+        'options.customEngineOptions': 'Op\u0163iuni motor analiz\u0103 personalizat',
+        'customEngineOptions.noCloud': 'Folose\u015fte numai motorul local pentru evalu\u0103ri',
+        'customEngineOptions.practice': 'Aplic\u0103 \u00een mod Practic\u0103',
+        'applyInPractice': 'Set\u0103ri motor personalizat \u00een mod Practic\u0103'
       }
     }
-
-    checkEngineLevel=()=>{
-      const parent=this.lichessTools;
-      const lichess=parent.lichess;
-      const analysis=lichess.analysis;
-      if (!analysis) return;
-      if (analysis.practice) return;
-      let customEngineDepth=+(parent.currentOptions.getValue('customEngineLevel'));
-      if (!customEngineDepth && this.options.noCloud) customEngineDepth=99;
-      const ceval = analysis.ceval;
-      if (!ceval.enabled()) return;
-      const state=ceval.getState();
-      const cloud=ceval.showingCloud();
-      const isIdle = state===2 || cloud;
-      const isRunning = state===3 && !cloud;
-      const node = analysis.node;
-      if (!ceval.lastStarted) {
-         return;
-      }
-      const step = ceval.lastStarted.steps.at(-1);
-      if (!step?.ceval) return;
-      const depth = step.ceval.depth;
-      if (isIdle) {
-	    if (((cloud && this.options.noCloud) || depth<customEngineDepth)) {
-          if (ceval.canGoDeeper()) {
-            node.autoDeeper=customEngineDepth;
-            node.startDepth=step.ceval.depth;
-            ceval.goDeeper();
-          }
-        } else {
-          if (ceval.isDeeper()) {
-            ceval.isDeeper(false);
-            analysis.redraw();
-          }
-        }
-      }
-      if (node.autoDeeper && depth>=customEngineDepth && (!cloud || !this.options.noCloud)) {
-        node.autoDeeper=null;
-        node.startDepth=null;
-        ceval.stop();
-        ceval.isDeeper(false);
-        analysis.redraw();
-      }
-      if (node.autoDeeper && node.autoDeeper!=customEngineDepth && ceval.enabled()) {
-        ceval.enabled(false);
-        ceval.enabled(true);
-        analysis.redraw();
-      }
-    };
-
-
-    injectCeval=()=>{
-      const parent=this.lichessTools;
-      const lichess=parent.lichess;
-      const analysis=lichess.analysis;
-      if (!analysis) return;
-      const ceval=analysis.ceval;
-      ceval.effectiveMaxDepth=parent.wrapFunction(ceval.effectiveMaxDepth,{
-        id:'customEngineLevel',
-        after:($this,oldMaxLevel)=>{
-          const node=analysis.node;
-          if (node.autoDeeper>(node.startDepth||oldMaxLevel)) {
-            return node.autoDeeper;
-          }
-          const customEngineDepth=+(parent.currentOptions.getValue('customEngineLevel'));
-          if (customEngineDepth && oldMaxLevel<customEngineDepth) {
-            return customEngineDepth;
-          } else {
-            return oldMaxLevel;
-          }
-        }
-      });
-    };
 
     analysisControls=()=>{
       const parent=this.lichessTools;
@@ -112,68 +45,173 @@
       const lichess=parent.lichess;
       const analysis=lichess.analysis;
       if (!analysis) return;
+
+      if (!analysis.practice) {
+        const customDepth=analysis.ceval?.infinite() || (analysis.ceval?.isDeeper() && !analysis.node.autoDeeper)
+          ? 99
+          : this.options.depth;
+        if (customDepth && analysis.ceval.enabled() && !analysis.ceval.showingCloud()) {
+          const elem=$('div.ceval div.engine span.info')[0];
+          if (elem) {
+            // lichess keeps a reference to the actual node
+            const textNode = Array.from(elem.childNodes).find(n=>n.nodeType==3);
+            if (textNode) {
+              const infoText=textNode.textContent;
+              const newText=infoText.replace(/(\d+)(?:\/\d+)?/,'$1/'+customDepth);
+              if (infoText!=newText) {
+                textNode.textContent=newText;
+              }
+            }
+          }
+        }
+      }
+
       $('.tview2').toggleClass('lichessTools-noCloud',this.options.noCloud);
       const container=$('div.analyse__tools div.action-menu');
       if (!container.length) return;
+
       if (!$('.abset-noCloud',container).length) {
-        const html=`<div class="setting abset-noCloud" title="LiChess Tools - $trans(options.noCloud)">
+        const html=`<div class="setting abset-noCloud" title="LiChess Tools - $trans(customEngineOptions.noCloud)">
       <div class="switch">
         <input id="abset-noCloud" class="cmn-toggle" type="checkbox" checked="">
         <label for="abset-noCloud"></label>
       </div>
-      <label for="abset-noCloud">$trans(options.noCloud)</label>
+      <label for="abset-noCloud">$trans(customEngineOptions.noCloud)</label>
     </div>`.replace(/\$trans\(([^\)]+)\)/g,m=>{
           return parent.htmlEncode(trans.noarg(m.slice(7,-1)));
         });
-        $(html).insertAfter($('div.abset-enable-nnue',container).eq(0));
+        $(html).insertAfter($('div.abset-gauge',container).eq(0));
         $('#abset-noCloud')
           .on('change',async ()=>{
             const options=parent.currentOptions;
-            options.noCloud=$('#abset-noCloud').is(':checked');
+            options.customEngineOptions=[$('#abset-noCloud').is(':checked')?'noCloud':'',this.options.practice?'practice':''].filter(o=>o).join(',');
             await parent.applyOptions(options);
             parent.fireReloadOptions();
           });
       }
       $('#abset-noCloud')
         .prop('checked',this.options.noCloud);
+
+
+      if (!$('.abset-practice',container).length) {
+        const html=`<div class="setting abset-practice" title="LiChess Tools - $trans(applyInPractice)">
+      <div class="switch">
+        <input id="abset-practice" class="cmn-toggle" type="checkbox" checked="">
+        <label for="abset-practice"></label>
+      </div>
+      <label for="abset-practice">$trans(applyInPractice)</label>
+    </div>`.replace(/\$trans\(([^\)]+)\)/g,m=>{
+          return parent.htmlEncode(trans.noarg(m.slice(7,-1)));
+        });
+        $(html).insertAfter($('div.abset-noCloud',container).eq(0));
+        $('#abset-practice')
+          .on('change',async ()=>{
+            const options=parent.currentOptions;
+            options.customEngineOptions=[this.options.noCloud?'noCloud':'',$('#abset-practice').is(':checked')?'practice':''].filter(o=>o).join(',');
+            await parent.applyOptions(options);
+            parent.fireReloadOptions();
+          });
+      }
+      $('#abset-practice')
+        .prop('checked',this.options.practice);
+
     };
 
     async start() {
       const parent=this.lichessTools;
       const value=+(parent.currentOptions.getValue('customEngineLevel'));
+      const customEngineOptions=parent.currentOptions.getValue('customEngineOptions');
       this.logOption('Custom engine level', value || 'Not set');
+      this.logOption('Custom engine options', customEngineOptions);
       this.options={
-        enabled: value,
-        noCloud: parent.currentOptions.getValue('noCloud'),
-        get isSet() { return this.enabled || this.noCloud; }
+        depth: value,
+        noCloud: parent.isOptionSet(customEngineOptions,'noCloud'),
+        practice: parent.isOptionSet(customEngineOptions,'practice'),
+        get isSet() { return this.depth || this.noCloud || this.practice; }
       };
-      this.logOption('Do not show cloud', this.options.noCloud);
       const lichess=parent.lichess;
       const analysis=lichess.analysis;
       if (!analysis) return;
-      parent.global.clearInterval(this.engineCheckInterval);
-      const ceval=analysis.ceval;
-      ceval.effectiveMaxDepth=parent.unwrapFunction(ceval.effectiveMaxDepth,'customEngineLevel');
-      if (this.options.isSet) {
-        this.injectCeval();
-        this.engineCheckInterval=parent.global.setInterval(this.checkEngineLevel,500);
-      } else {
-        if (ceval.enabled()) {
-          ceval.enabled(false);
-          ceval.enabled(true);
-          analysis.redraw();
-        }
-      }
+
       lichess.pubsub.off('redraw',this.analysisControls);
-      lichess.analysis.actionMenu.toggle=lichessTools.unwrapFunction(lichess.analysis.actionMenu.toggle,'noCloud');
+      lichess.analysis.actionMenu.toggle=lichessTools.unwrapFunction(lichess.analysis.actionMenu.toggle,'customEngineOptions');
       lichess.pubsub.on('redraw',this.analysisControls);
       lichess.analysis.actionMenu.toggle=lichessTools.wrapFunction(lichess.analysis.actionMenu.toggle,{
-        id:'noCloud',
+        id:'customEngineOptions',
         after: ($this, result, ...args)=>{
           parent.global.setTimeout(this.analysisControls,100);
         }
       });
       this.analysisControls();
+
+      const determineCevalState=()=>{
+        if (analysis.practice?.running() && !this.options.practice) return;
+        const node=analysis.ceval.lastStarted?.steps?.at(-1);
+        const curDepth=node?.ceval?.depth;
+        if (analysis.ceval.enabled() && analysis.ceval.showingCloud() && analysis.ceval.canGoDeeper()
+            && (this.options.noCloud || (this.options.depth && curDepth!==undefined && curDepth<this.options.depth)))
+        {
+          node.autoDeeper=this.options.depth;
+          analysis.ceval.goDeeper();
+          analysis.redraw();
+        }
+        if (analysis.ceval.enabled() && !analysis.ceval.showingCloud() && node.autoDeeper
+            && this.options.depth && curDepth!==undefined && curDepth>=this.options.depth)
+        {
+          node.autoDeeper=undefined;
+          analysis.ceval.stop();
+          analysis.redraw();
+        }
+      };
+
+      if (this.options.depth||this.options.noCloud) {
+        if (!parent.isWrappedFunction(analysis.evalCache.onCeval,'customEngineOptions')) {
+          analysis.evalCache.onCeval=parent.wrapFunction(analysis.evalCache.onCeval,{
+            id:'customEngineOptions',
+            before:($this,...args)=>{
+              determineCevalState();
+            },
+            after:($this,result,...args)=>{
+              ;
+            }
+          });
+        }
+      } else {
+          analysis.evalCache.onCeval=parent.unwrapFunction(analysis.evalCache.onCeval,'customEngineOptions');
+      }
+
+      if (this.options.noCloud) {
+        if (!parent.isWrappedFunction(analysis.evalCache.fetch,'customEngineOptions')) {
+          analysis.evalCache.fetch=parent.wrapFunction(analysis.evalCache.fetch,{
+            id:'customEngineOptions',
+            before:($this,...args)=>{
+               if (!this.options.noCloud) return;
+               if (analysis.practice?.running() && !this.options.practice) return;
+               return false;
+            }
+          });
+        }
+        if (!parent.isWrappedFunction(analysis.explorer.fetchTablebaseHit,'customEngineOptions')) {
+          analysis.explorer.fetchTablebaseHit=parent.wrapFunction(analysis.explorer.fetchTablebaseHit,{
+            id:'customEngineOptions',
+            before:($this,...args)=>{
+               if (!this.options.noCloud) return;
+               if (analysis.practice?.running() && !this.options.practice) return;
+               return false;
+            },
+            after:($this,result,...args)=>{
+               if (!this.options.noCloud) return;
+               if (analysis.practice?.running() && !this.options.practice) return;
+               return Promise.reject('Cloud disabled');
+            }
+          });
+        }
+      } else {
+        analysis.evalCache.fetch=parent.unwrapFunction(analysis.evalCache.fetch,'customEngineOptions');
+        analysis.explorer.fetchTablebaseHit=parent.unwrapFunction(analysis.explorer.fetchTablebaseHit,'customEngineOptions');
+      }
+
+      determineCevalState();
     }
 
   }